studytrack/backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-pptx
supabase
python-dotenv
httpx
